requirements.txt
streamlit
python-dotenv
google-genai
streamlit-cookies-manager